chore: Move fixtures to conftest
